working on having somewhat deterministic applying of force. This commit seems to get this working as I'm using the bullet callback for applying the forces. The custom callback class I made here would then be used for applying the torques from my simulation, and I guess, also run the kernels.

+++ multileg/src/util/UniqueIndexList.h
@@ -50,7 +50,7 @@ unsigned int UniqueIndexList<T>::add(T p_valueRef)
 	else
 	{
 		m_list.push_back(p_valueRef);
-		index = m_list.size()-1;
+		index = (unsigned int)m_list.size()-1;
 	}
 	return index;
 }
